fix shadowed names
Joe, your build errors are because you are shadowing names `head` and `tail` that are already defined in the Haskell Prelude. If you look at hw01.cabal, you will see the `-Wall` flag is passed to `ghc-optiond`. This means that all warnings (shadowed name, in this case) become errors. Just rename these so that they don't conflict.